requirements.txt
paho-mqtt
PyQt5
pandas
icecream
